#14: add configuration model and rename folders to singular
