+++ lib/bloc/home/home_event.dart
@@ -1,0 +1,1 @@
+sealed class HomeEvent {}
