+++ lib/model/currency.dart
