+++ lib/model/membership.dart
