+++ lib/repository/exchange_rate_repository/exchange_rate_repository.dart
@@ -1,5 +1,5 @@
-import "package:currency_converter/models/currency.dart";
-import "package:currency_converter/models/exchange_rate.dart";
+import "package:currency_converter/model/currency.dart";
+import "package:currency_converter/model/exchange_rate.dart";
 
 abstract class ExchangeRateRepository {
   Future<ExchangeRateSnapshot> getExchangeRateSnapshot(CurrencyCode currencyCode);
