+++ lib/repository/exchange_rate_repository/local_exchange_rate_repository.dart
@@ -1,7 +1,7 @@
 import "dart:math";
 
-import "package:currency_converter/models/currency.dart";
-import "package:currency_converter/models/exchange_rate.dart";
+import "package:currency_converter/model/currency.dart";
+import "package:currency_converter/model/exchange_rate.dart";
 import "package:currency_converter/repository/exchange_rate_repository/exchange_rate_repository.dart";
 import "package:currency_converter/repository/local_repository_config.dart";
 import "package:currency_converter/util/errors/cc_error.dart";
