+++ pubspec.yaml
@@ -67,7 +67,6 @@ dev_dependencies:
 
   flutter_launcher_icons: ^0.13.1
   mocktail: ^1.0.4
-  drift_dev: ^2.19.1
   build_runner: ^2.4.12
   realm_common: ^20.0.0
