+++ test/repository/exchange_rate_repository/github_exchange_rate_repository_test.dart
@@ -1,4 +1,4 @@
-import "package:currency_converter/models/currency.dart";
+import "package:currency_converter/model/currency.dart";
 import "package:currency_converter/repository/exchange_rate_repository/github_exchange_rate_repository.dart";
 import "package:currency_converter/util/errors/cc_error.dart";
 import "package:flutter_test/flutter_test.dart";
