+++ test/repository/membership_repository/local_membership_repository_test.dart
@@ -1,4 +1,4 @@
-import "package:currency_converter/models/membership.dart";
+import "package:currency_converter/model/membership.dart";
 import "package:currency_converter/repository/membership_repository/local_membership_repository.dart";
 import "package:currency_converter/util/errors/cc_error.dart";
 import "package:flutter_test/flutter_test.dart";
